Updated Read Me
Outlined the project overview, 2 possible approaches to tackling the project and what each approach would mean for the code base and work load. Highlighted the challenges involved as well as the approach taken. As a last minute addition, ajax was added to the project to use the tool without repeatedly requiring a form submission.