Added project files via upload
Added the 'completed' project to the repository. v1.0.0.0 did not use ajax, though v1.0.0.1 does. Ajax is used to that on input to the form, a new FormData object is created and sent off to the server. Upon success, the response is entered into the to_value input. This addition allows the end-user to use it without having to press submit endlessly.