Add files via upload
Made several changes locally. Created a new file, /includes/constants.php, as well as /includes/utility-functions.php which hold script constants and utility functions, respectively. Also created /includes/shared/zero.txt which contains a mock response by the server. Within the /assets/js/script.js file, an ajax call is made to the server to process a form submission on input. zero.txt is used in a get request when the form input is deleted and its value is equal to an empty string.