Update script.js
Moved most of the variables within a global scope to the top of the page. Defaulted theToValue, theFromValue, theToSelect, and theFromSelect variables to an actual value to prevent a TypeError when theToSelect was changed without typing in the input first. Added an anonymous function to select text in each input on focus.  Added in a jQuery plugin $.QueryString to grab current url params and created a function called select. The final two additions allowed for the script to 'know' which page the user is viewing. As such, it now selects default values -- outlined by the client coder -- from the select lists as well as a default value for the to_value_input. Once the default values are selected and placed, the form is submitted providing the end-user some default info instead of seeing just an empty form that needs to be completed.