Update temperature.class.php
updated the method name that processes the conversion from convertSpeed() to processConversion to match each of the other classes for uniformity. This file still contains the previous convertFromCelsius and convertToCelsius methods as, internally, it operates a bit differently for calculations. This is why it does not utilize the \calc\Calculate trait.